docker/tensorflow-lite-micro-rtos-fvp/requirements.txt
ethos-u-vela==3.0.0
pillow==8.2.0
pygame==2.0.1
PySimpleGui==4.41.2 
requests==2.25.1
pyelftools==0.27
packaging==20.9